refactor: 💡 move petFilters file and update imports
