+++ src/components/SearchPetForm.tsx
@@ -4,7 +4,7 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import Select from 'react-select'
 import { api } from '../lib/axios'
 import usePetsContext from '../hooks/usePetsContext'
-import { PetGender, PetType } from '../utils/petFilters'
+import { PetGender, PetType } from '../constants/petFilters'
 import { customStyles } from '../styles/selectStyles'
 import { Response } from '../context/petsContext'
 import { SearchPetFormData, searchPetSchema } from '../@types/zodTypesSearchPet'
